add mod check and add list commands command

+++ src/cogs/custom_command.py
@@ -10,6 +10,9 @@ class CustomCommandCog(commands.Cog):
 
     @commands.command(aliases=["addcmd"])
     async def add_command(self, ctx, label: str, response: str):
+        if not ctx.author.is_mod:
+            return
+
         self.db.add_custom_command(
             streamer=ctx.channel.name, label=label, response=response
         )
@@ -17,8 +20,18 @@ class CustomCommandCog(commands.Cog):
 
     @commands.command(aliases=["rmcmd"])
     async def remove_command(self, ctx, label: str):
+        if not ctx.author.is_mod:
+            return
         self.db.remove_custom_command(streamer=ctx.channel.name, label=label)
         await ctx.send("Command removed, maybe.")
+
+    @commands.command(aliases=["lscmd"])
+    async def list_commands(self, ctx):
+        commands = self.db.get_streamer_custom_commands(streamer=ctx.channel.name)
+
+        names = [command.get("label") for command in commands]
+
+        await ctx.send(f"Available commands: {', '.join(names)}")
 
 
 def prepare(bot: commands.Bot):
